+++ assets/javascripts/kitten/components/molecules/buttons/button/button.stories.js
@@ -1,12 +1,17 @@
 import React from 'react'
-import { Button } from './index'
+import {
+  Button,
+  buttonModifiers,
+  buttonFitOptions,
+  buttonMobileFitOptions,
+  buttonSizes,
+  buttonVariants,
+} from './index'
 import { HeartIcon, PayPalIcon, Badge, Cart } from '../../../..'
 import { DocsPage } from 'storybook/docs-page'
 
 const args = {
   borderRadius: 0,
-  fluid: false,
-  icon: false,
   modifier: 'hydrogen',
   rounded: false,
   disabled: false,
@@ -14,6 +19,8 @@ const args = {
   tag: undefined,
   variant: 'orion',
   children: 'My Button',
+  fit: 'min-width',
+  mobileFit: undefined,
 }
 
 const argTypes = {
@@ -21,36 +28,9 @@ const argTypes = {
     name: 'borderRadius',
     control: 'number',
   },
-  fluid: {
-    name: 'fluid',
-    control: 'boolean',
-  },
-  icon: {
-    name: 'icon',
-    control: 'boolean',
-  },
   modifier: {
     name: 'modifier',
-    options: [
-      'hydrogen',
-      'helium',
-      'lithium',
-      'beryllium',
-      'carbon',
-      'oxygen',
-      'checked',
-      'copper',
-      'boron',
-      'neon',
-      'iron',
-      'calcium',
-      'social_facebook',
-      'social_twitter',
-      'social_linkedin',
-      'social_instagram',
-      'social_youtube',
-      'social_pinterest',
-    ],
+    options: buttonModifiers,
     control: 'select',
   },
   rounded: {
@@ -59,7 +39,7 @@ const argTypes = {
   },
   size: {
     name: 'size',
-    options: ['nano', 'micro', 'tiny', 'regular', 'big', 'huge', 'giant'],
+    options: buttonSizes,
     control: 'select',
   },
   tag: {
@@ -68,12 +48,22 @@ const argTypes = {
   },
   variant: {
     name: 'variant',
-    options: ['andromeda', 'orion'],
+    options: buttonVariants,
     control: 'inline-radio',
   },
   children: {
     name: 'children',
     control: 'text',
+  },
+  fit: {
+    name: 'fit',
+    control: 'select',
+    options: buttonFitOptions,
+  },
+  mobileFit: {
+    name: 'mobileFit',
+    control: 'select',
+    options: buttonMobileFitOptions,
   },
 }
 
@@ -105,13 +95,13 @@ export default {
 export const Default = args => <Button {...args} />
 
 export const WithIcon = args => (
-  <Button {...args} icon>
+  <Button {...args} fit="icon">
     <HeartIcon width={15} height={15} />
   </Button>
 )
 
 export const WithColorIcon = args => (
-  <Button {...args} icon>
+  <Button {...args} fit="icon">
     <PayPalIcon />
   </Button>
 )