+++ assets/javascripts/kitten/components/molecules/buttons/close-button/index.js
@@ -8,7 +8,7 @@ export const CloseButton = ({
   ...others
 }) => (
   <Button
-    icon
+    fit="icon"
     type="button"
     title={closeButtonLabel}
     aria-label={closeButtonLabel}