+++ assets/javascripts/kitten/karl/pages/news-list/index.js
@@ -107,7 +107,7 @@ const NewsList = () => {
       </Head>
       <Container>
         <NewsContainer>
-          <Button big fluid modifier="oxygen">
+          <Button size="big" fit="fluid" modifier="oxygen">
             <InstrumentTagIcon width="14" />
             <span>Créer une nouvelle actu</span>
           </Button>